code/requirements.txt
boto3==1.40.28
botocore==1.40.28
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
docopt==0.6.2
duckdb==1.3.2
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.23
pymgclient==1.5.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyproj==3.7.2
pystomp==0.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.14.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
stomp.py==8.2.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.198
ultralytics-thop==2.0.17
urllib3==2.5.0
websocket-client==1.8.0
